+++ src/EasyCodingStandard/Factory.php
@@ -1,12 +1,5 @@
 <?php declare(strict_types=1);
 
-/**
- * Copyright (C) Brian Faust
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
 namespace Cline\CodingStandard\EasyCodingStandard;
 
 use Cline\CodingStandard\PhpCsFixer\CopyrightHeader;
@@ -22,6 +15,7 @@ use Cline\CodingStandard\PhpCsFixer\Fixer\PhpdocLineLengthFixer;
 use Cline\CodingStandard\PhpCsFixer\Fixer\PsalmImmutableOnReadonlyClassFixer;
 use Cline\CodingStandard\PhpCsFixer\Fixer\RedundantReadonlyPropertyFixer;
 use Cline\CodingStandard\PhpCsFixer\Fixer\RemoveAuthorTagFixer;
+use Cline\CodingStandard\PhpCsFixer\Fixer\RemoveHeaderCommentFixer;
 use Cline\CodingStandard\PhpCsFixer\Fixer\RemoveVersionTagFixer;
 use Cline\CodingStandard\PhpCsFixer\Preset\PresetInterface;
 use Cline\CodingStandard\PhpCsFixer\Preset\Standard;
@@ -145,6 +139,7 @@ final class Factory
             new PsalmImmutableOnReadonlyClassFixer(),
             new RedundantReadonlyPropertyFixer(),
             new RemoveAuthorTagFixer(),
+            new RemoveHeaderCommentFixer(),
             new RemoveVersionTagFixer(),
         ];
     }