+++ src/PhpCsFixer/CopyrightHeader.php
@@ -1,12 +1,5 @@
 <?php declare(strict_types=1);
 
-/**
- * Copyright (C) Brian Faust
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
 namespace Cline\CodingStandard\PhpCsFixer;
 
 /**